+++ packages/math-input/src/types.ts
@@ -1,3 +1,5 @@
+import ReactDOM from "react-dom";
+
 import {CursorContext} from "./components/input/cursor-contexts";
 import Key from "./data/keys";
 import {
@@ -85,3 +87,12 @@ export type ActiveNodesObj = {
 export type LayoutProps = {initialBounds: Bound};
 
 export type ClickKeyCallback = (key: Key) => void;
+
+export interface KeypadAPI {
+    activate: () => void;
+    dismiss: () => void;
+    configure: (configuration: KeypadConfiguration, cb: () => void) => void;
+    setCursor: (cursor: Cursor) => void;
+    setKeyHandler: (keyHandler: KeyHandler) => void;
+    getDOMNode: () => ReturnType<typeof ReactDOM.findDOMNode>;
+}